requirements.txt
Flask==2.3.2
pandas==2.0.3
scikit-learn==1.3.2
numpy>=1.21
requests==2.32.3
python-dotenv==1.0.1
Flask-Caching==2.3.0
mlflow==2.11.3
redis==5.0.4